move remaining RangeIndex arithmetic tests
